+++ include/mps/cu_test_helpers.h
@@ -1,3 +1,3 @@
 #pragma once
 
-void test_cal_Indepl0(const float *C, const int *M, const int *P, const int *W, int *G, const double *Th);
+void test_cal_Indepl0(const float *C, const int *M, const int *P, const int *W, int *G, const float *Th);
